Extends the create procedure of persistence/varnish node to wake stopped ones
The previous implementation did not respect the stoped state of those instances.
Instances were not woking up, and that was breaking the deployment.# With this change, machines that are stopped will be booted automatically.

Improvement indirectly fixes the problem when the node was in other than the preferred availability zone. Currently, the node will be used. In the past, a new machine was created.