[IMPROVEMENT] Paramerize and adjust app node nginx fastcgi buffer size configuration
The most imporatant change is the increase of fastcgi_buffer_size parameters which
if set too low will lead to 'too big header' errors, which seem to be relatively
frequent with Magento.

This changes should not have much effect on memory consumption as the total buffer
size does not change. The number of buffers and size of single one have been swapped
which should decrease memory fragmentation and maybe even slightly improve performance
at a negligible (in comparison to PHP memory usage) memory usage increase.