[IMPROVEMENT] Improve zram/vmtouch/swap handling
Since the php-cli-opcache contained improvements to the vmtouch/preload
mechanism merging it had the consequence of reenabling the vmoutch by
default which is not desireable anymore.

A more fine-grained switches for zram/fscache-preload/kernel-vm/swap
parameters have been implemented with default "smart" mechanism which
should enable/disable those feature based on the parameters of the
target node.