[IMPROVEMENT] Proper rolling AutoScalingGroup instance refresh
Since the Ansible module for this is very buggy and basically tries to perform this
whole logic "in software" by polling for the states of multiple partial operations.

This PR replaces the broken and unmaintained ec2_asg module's instance refresh
mechanism, instead performing it directly using AWS API calls (with the help of aws-cli).

There is basically one API call needed and it does the whole job, start to finish perfectly
without any intervention.

After refresh is started the play just waits until the operation reports 100% completion
status in order to know when the instances are ready which works way better than before.

The whole thing seems to be working perfectly.